api-gateway/requirements.txt
Flask==1.0.2
requests
paho-mqtt
flask-mqtt
sortedcontainers
